25-02-11 모델 메모리 업데이트, 모델 topic 대화 작업중
