+++ app/templates/base.html
@@ -8,7 +8,7 @@
     <link rel="stylesheet" href="{{ url_for('static', filename='css/bootstrap.min.css') }}">
     <!-- styles CSS -->
     <link rel="stylesheet" href="{{ url_for('static', filename='css/styles.css') }}">
-    <title>🔮Tarot Nana🔮</title>
+    <title>🔮 Tarot Nana</title>
 </head>
 <body>
 {% block content %}
