+++ app/templates/tarot_chat copy 2.html
@@ -10,5 +10,4 @@
 </div>
 <!-- JavaScript 연결 -->
 <script src="{{ url_for('static', filename='js/chat.js') }}"></script>
-
 {% endblock %}
