--- app/templates/tarot_chat copy.html
@@ -1,73 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>타로 챗봇</title>
-    <style>
-        body { font-family: Arial, sans-serif; text-align: center; }
-        #chat-container { width: 400px; margin: 20px auto; border: 1px solid #ccc; padding: 10px; border-radius: 5px; }
-        .chat-box { height: 300px; overflow-y: auto; border-bottom: 1px solid #ddd; padding-bottom: 10px; }
-        .message { padding: 8px; margin: 5px; border-radius: 5px; display: inline-block; max-width: 80%; }
-        .user { background: #d1e7dd; text-align: right; float: right; clear: both; }
-        .bot { background: #f8d7da; text-align: left; float: left; clear: both; }
-        input, button { margin-top: 10px; padding: 8px; width: 80%; }
-    </style>
-</head>
-<body>
-    <h2>타로 챗봇</h2>
-    <div id="chat-container">
-        <div class="chat-box" id="chat-box"></div>
-        <input type="text" id="user-input" placeholder="메시지를 입력하세요..." />
-        <button onclick="sendMessage()">보내기</button>
-    </div>
-
-    <script>
-        async function sendMessage() {
-            let userInput = document.getElementById("user-input");
-            let chatBox = document.getElementById("chat-box");
-            let message = userInput.value.trim();
-            if (!message) return;
-        
-            // ✅ 사용자 입력을 화면에 추가
-            chatBox.innerHTML += `<div class="message user">${message}</div>`;
-            userInput.value = "";
-        
-            // ✅ 말풍선 생성 (비어 있는 채로 만들어 놓음)
-            let botMessage = document.createElement("div");
-            botMessage.classList.add("message", "bot");
-            chatBox.appendChild(botMessage);
-        
-            try {
-                // ✅ Flask에 POST 요청
-                let response = await fetch("/chat/chat", {
-                    method: "POST",
-                    headers: { "Content-Type": "application/json" },
-                    body: JSON.stringify({ message })
-                });
-        
-                // ✅ chunk 단위로 응답을 받아서 말풍선 내부에 추가
-                const reader = response.body.getReader();
-                let decoder = new TextDecoder();
-                let fullResponse = "";  // ✅ 전체 응답을 저장할 변수
-        
-                async function readChunks() {
-                    let { done, value } = await reader.read();
-                    while (!done) {
-                        let chunkText = decoder.decode(value, { stream: true });
-                        fullResponse += chunkText;  // ✅ 기존 말풍선 안에 계속 추가
-                        botMessage.innerHTML = fullResponse;  // ✅ 말풍선 내부 텍스트 갱신
-                        chatBox.scrollTop = chatBox.scrollHeight;
-                        ({ done, value } = await reader.read());
-                    }
-                }
-        
-                readChunks();
-            } catch (error) {
-                console.error("Error:", error);
-            }
-        }
-    </script>
-        
-</body>
-</html>
